scripts/requirements-scrape.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
python-frontmatter>=1.0.0
